fix: LEAP-248: Fix app crashing by hotkeys  (#1653)
* fix: LEAP-248: Fix app crashing by multiple hotkeys

The 'unbind' function in Hotkey.ts has been enhanced to handle multiple keys provided as a string. A new helper function 'getKeys' was added to split the key string and return an array of keys. Each key in the array is then processed individually in the 'unbind' function. This helps keymaster to handle more that one hotkey.

* Add alias for comma to prevent keymaster fails

Added a new mapping for the comma key in the ALIASES object. Also, refactored the applyAliases function to use the getKeys function for better tokenization and handling of keys. This change helps in fixing the keymaster inconsistencies with certain keys.

* Add regression tests for hotkeys problem

* Fix import paths

* Improve the comment

---------

Co-authored-by: hlomzik <hlomzik@users.noreply.github.com>